requirements.txt
attrs==25.3.0
Automat==25.4.16
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.12.2
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.3.0
cloudscraper==1.2.71
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.7
cssselect==1.3.0
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
et_xmlfile==2.0.0
fake-useragent==2.2.0
filelock==3.19.1
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
h11==0.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.12.2
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
loguru==0.7.3
lxml==5.3.0
MarkupSafe==3.0.3
mysql-connector-python==9.4.0
mysqlclient==2.2.7
numpy==2.3.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
parsel==1.10.0
platformdirs==4.4.0
Protego==0.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyDispatcher==2.0.7
PyMySQL==1.1.2
pyOpenSSL==25.1.0
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
queuelib==1.8.0
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
scrapfly-sdk==0.8.23
Scrapy==2.13.3
selenium==4.11.2
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.44
tldextract==5.3.0
trio==0.30.0
trio-websocket==0.12.2
Twisted==25.5.0
typing_extensions==4.14.1
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.0.3
user-agents==2.2.0
virtualenv==20.34.0
w3lib==2.3.1
webdriver-manager==4.0.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
win32_setctime==1.2.0
wsproto==1.2.0
zope.interface==8.0
